dialog: Remove no-op SingleChildScrollView on iOS
In #1017, we overlooked the fact that a SingleChildScrollView is
added automatically on iOS but not on Android.

I haven't reproduced an observable bug that comes from this, but it
calls for a fix.

I tested this change manually on iOS (showErrorDialog,
showSuggestedActionDialog, and UpgradeWelcomeDialog), with short
text and long text (longer than a screenful, to check the scrolling
works).