user [nfc]: Expose UserStoreImpl.userMapFromInitialSnapshot
We'll need to pass RealmStoreImpl some details of the self-user, in
order to correctly interpret (for #814) the permissions that live
on RealmStore.

But because UserStore already depends on RealmStore, we don't want
to introduce a dependency in RealmStoreImpl on UserStore as a whole.
So separate out the step that processes the user lists, and do that
step in advance before constructing RealmStoreImpl.