content: Remove TextStyle merge in rendering inline code
Also in rendering inline TeX source when it doesn't parse...but
that's decreasingly visible as we implement more TeX. :)

Fixes #806.
Fixes #1812.

Similar reasoning as in the previous commits; see there. The
inline-code span-node was clobbering the link-color and font-weight
attributes when those are set by ancestor link and emphasis nodes.