group: Add selfInGroupSetting
This is the way we'll be inspecting group membership for checking
permissions (#814), which is the main way we'll be using group
membership in general.

Because this method is how most of the app will consume group
membership, adding this method also lets us write natural tests for
the logic that maintains group membership on events.

At this stage we stick with a very straightforward implementation.
We might optimize it later with more data structures.