content: Remove TextStyle merge in rendering unicode emoji
Fixes #1818.

The InlineContent widget creates a Text.rich with a tree of
`InlineSpan`s, and the widget's `style` param is passed to the root
span in that tree. That style's attributes will propagate down the
tree as defaults, which means that a non-root node will still be
styled with e.g. h1 font size if it doesn't `TextStyle.merge` that
root style into its own style, as was happening here.

That merging can also be unhelpful, not just unnecessary, because it
can clobber attributes that were set at nodes between the current
node and the root. This kind of clobbering causes some known issues;
the first of these is fixed in this commit, and we'll fix the rest,
coming up:
- #1818 content: Unicode emoji in strikethrough should have
  strikethrough line, but doesn't
- #1817 content: Bold text in strikethrough should have
  strikethrough line, but doesn't
- #1812 content: Inline code in a bold span should be bold (but
  isn't)
- #806 content: Inline code links should be rendered with the link
  color

We'll fix the others, coming up.