content: Include inline styles (italic etc.) in what we pass to WidgetSpans
Fixes #1813.
Fixes #1819.

This is NFC for inline math, since all the inline styles are
overridden there. But see the previous commit for where we
consolidated those overrides, making a natural place to add more if
needed in the future, with TODOs for some existing ones we might not
want to keep.