emoji_reaction test: Use pumpAndSettle to let overscroll animation finish
These tests scroll a long way, in order to be sure to scroll to the
end of the list.  That means they invoke overscroll.  The details of
how overscroll behaves are changing in an upstream PR:
  https://github.com/flutter/flutter/pull/173849

The new behavior will take more than a single `pump` to reach the
end state where the subsequent checks behave as expected.  These
tests aren't about the overscroll behavior, so just gloss over any
wrinkles by calling `pumpAndSettle`.