katex [nfc]: Rename a local to hasVlistT2
This reflects the condition that's directly encoded by this boolean.
(Through several conditionals in the rest of the function, we end up
confirming that this value is true just if the element does in fact
have two `.vlist-r` children, as the old name suggests.  But that
comes later and more indirectly.)