button: Implement a disabled state for Toggle
Vlad suggested using 0.5 opacity if no special disabled state is
provided:
  https://chat.zulip.org/#narrow/channel/530-mobile-design/topic/toggle.3A.20disabled.20state/near/2250883
but Greg and I prefer 0.4 in this case to make it more distinct.

We may or may not end up using this for unsubscribed channels
without content access (i.e. ones that you can't subscribe yourself
to). We may instead just not show a toggle for those channels.