deps: Upgrade Flutter to work around Flutter version-solving bug
CI is broken by what looks like a bug in `flutter pub get`:

  Resolving dependencies...
  The current Flutter SDK version is 3.37.0-1.0.pre-175.

  Because zulip requires Flutter SDK version >=3.37.0-1.0.pre-89,
  version solving failed.
  Failed to update packages.

It looks like the number at the end is being compared
lexicographically, not numerically as it should be. That should be
fixed, but for now, work around by just bumping the threshold to a
version where that number starts with "1".