tools/bump-version: Better centralize version format; handle mismatch
This is NFC except in its handling of errors.

I haven't been using this script since the app launched (and have
instead done its work manually), because it assumes the version
numbers are 0.0.N+N.  After this commit, it's a small local change to
make the script handle the new 30.0.N+N version numbers, and the
script will also fail cleanly if we ever change the format again.