tools/bump-version: Update for current version numbers
In principle it wouldn't be hard to generalize this logic further,
reading out the current major and minor version numbers rather than
expecting them to have certain values.

But it's not obvious what direction we'd want such generalization
to go in -- that depends on what our version-numbering practices
would be after they change.  On current plans, the versions will be
30.0.N+N indefinitely, so it's inherently uncertain what pattern we
might choose (and therefore how the version should get bumped by
this script) after we change those plans, if we someday do.
So just leave the generalization to be done if and when we need it.