tools/format-changelog: Upgrade line-unwrapping to handle nested lists
The `fmt` tool from GNU coreutils is very simple, and doesn't have the
needed configuration knobs to be able to handle the nested lists that
we now routinely use in changelog entries; it ends up mangling them.

As a result, for many recent releases I've been doing the reformatting
semi-manually, in Emacs: M-q to unwrap lines (after setting the text
width to 999), and then a regexp find-and-replace to turn references
`#1234` into `#F1234`.

It seems to be somewhat annoying to get Emacs to operate within a
shell pipeline.  But it's less annoying for Vim; and Vim's `gq`
works just as well as Emacs's M-q, given the right config.
So use that.