db [nfc]: Ignore upcoming experimental_member_use warning from analyzer
In an upcoming Dart SDK change
(https://dart-review.googlesource.com/c/sdk/+/450970), I intend to add
logic to the analyzer for generating a warning if an API marked
@experimental is used. This will allow experimental analyzer features
to be developed without creating a risk of breaking changes
downstream.

To avoid this change producing a failure in the customer_testing bot
(https://ci.chromium.org/ui/p/flutter/builders/try/Linux%20customer_testing),
the `zulip-flutter` customer test needs to first be modified in order
to ignore the upcoming warning.

[greg: added comment]