unread_count_badge: Remove `bold` variant; pin down to current Figma
The "Channels" page and "Inbox" page both have updated designs in
Figma, which we should follow.

This commit matches the badges on the "Channels" page to that new
design. There's more work to match the badges on the "Inbox" page,
but now that work is defined crisply in a TODO:

  // TODO support the "kind=quantity" variant, update dartdoc

Related: #1406
Related: #1527