api: Allow SubscriptionUpdateEvent with unrecognized property
Our event poll shouldn't crash on subscription/update events about
data that we don't store and act on. This bugfix allows servers to
add new Subscription fields and send corresponding events without
causing the app to discard its data and reregister a new event
queue.

Soon, for #1837, we'd like to remove
SubscriptionProperty.inHomeView, which was deprecated in FL 139.
Events with in_home_view are still sent by modern servers (as of CZO
2025-10-03), but we never need them because they always come with an
is_muted event. Now, when we remove SubscriptionProperty.inHomeView,
the in_home_view event will parse instead of not parsing, and it'll
have SubscriptionProperty.unknown for `property`, which the store
already ignores.