android: Lower minSdkVersion to 26, aka Android 8 Oreo
Fixes #1898.

In general, our policy for supporting old OS versions is
described here (in the legacy app's repo, but it applies equally
to this app):
  https://github.com/zulip/zulip-mobile/blob/eb8505c4a/docs/architecture/platform-versions.md

So for versions old enough that they represent a very small fraction
of our overall user base, we don't have the resources to actively
maintain support; but if the app works on those versions, we leave
them supported in the app's metadata.

When we increased the minimum to Android 9, the reason was that we
weren't sure whether things would break on older Android versions,
and just didn't have a device older than Android 9 to test on:
  https://chat.zulip.org/#narrow/channel/243-mobile-team/topic/flutter.3A.20Java.20versions.3F/near/1775991

But in #1898 we've learned that the app does in fact work on
Android 8.  So restore that support in the app's manifest.