content test: Add explicit test case for Zulip Cloud external image preview
This kind of URL is already covered in the tests of *clusters* of
image previews, but it seems helpful to isolate it like this with
its own comment.

We're also about to add a third variant, testing that we accept
images generated with an arbitrary CAMO_URI, which actually we don't
currently do, so that'll be a regression test.