content: Parse `data-animated` on image-preview HTML
This might've been done more simply by just adding the boolean value
as a new field like `thumbnailAnimated` on ImagePreviewNode.

But this value works closely with some structured data in thumbnail
URL paths, here in image previews and also in the upcoming "inline
images" feature, #1913. So this new ImageThumbnailLocator class
seemed like a helpful way to present the data to consumers.