msglist test: Switch to using eg.defaultStreamMessageStreamId, where boring
The `prepare` helper, and now the prepare-outbox helpers, fall back
to this channel ID when a different one isn't specified. That's
sensible and lets these tests be a bit more compact.

When tests use two different channels, with pairs of variables like
`stream` and `otherStream`, or `stream1` and `stream2`, I let that
be, even though it might be reasonable for some of them to use
eg.defaultStreamMessageStreamId for one of the channels.