msglist test: Switch some model tests to use subscribed channels
Before this, in all the tests touched here, we were exercising the
message list with a channel that isn't actually in the
PerAccountStore. When we don't pass `stream` to the `prepare`
function, the function adds a *different* channel to the store, with
a corresponding Subscription, having streamId
eg.defaultStreamMessageStreamId.

The normal, boring setup, which is suitable for all these tests, is
to exercise the message list in a known, subscribed channel. Provide
that setup by passing the same `stream` object to `prepare` that we
use to create messages and narrows.