msglist test [nfc]: Fix some test descriptions
In all of these tests, the message list's narrow is a ChannelNarrow,
and all of the deleted messages are in that channel. The tests
aren't about whether or not the deleted messages are in the message
list's narrow; really they're about whether or not the deleted
messages are present in the MessageListView model.