deps: Update http to 1.6.0, from 1.5.0
Changelog:
  https://pub.dev/packages/http/changelog#160

One breaking change: in the Content-Type header, the `charset`
parameter is no longer set automatically, except on certain types
which require it.
  https://github.com/dart-lang/http/blob/a22386e9c/pkgs/http/lib/src/request.dart#L18-L31
  https://github.com/dart-lang/http/pull/1798

That breaks some of our tests which expect to see `charset=utf-8`,
so this commit includes a fix for those tests, simply removing that
expectation.  This is OK because the content type in question is
application/x-www-form-urlencoded, and that content type does not
define `charset` anyway:
  https://www.iana.org/assignments/media-types/application/x-www-form-urlencoded
In fact, it requires the use of UTF-8:
  https://url.spec.whatwg.org/#application/x-www-form-urlencoded

The package's default encoding for actually producing the body
remains UTF-8:
  https://github.com/dart-lang/http/blob/a22386e9c/pkgs/http/lib/src/request.dart#L179

[greg: expanded commit message]