content: Add ImageThumbnailLocator, wrapping thumbnail URLs
Soon we'll add a field on this for the `data-animated` attribute,
then write some logic to help UI code choose a thumbnail format from
RealmStore.serverThumbnailFormats.

Not NFC just because we've added a restriction on `src` for a
thumbnail URL in an image preview: now it must go through Uri.parse
without error, else we'll show an "unimplemented content" block.