compose: Show the Refresh/Subscribe banner when can send, too
Fixes #1798.

In #1873, we started showing a Refresh/Subscribe banner when viewing
an unsubscribed channel -- but only if you didn't have permission to
send messages there. We postponed showing it in the case where you
*can* send messages because the UI message for that case is "Replies
to your messages will not appear automatically", and that message
would have been confusing in a world where your own messages didn't
appear automatically either because of bugs that hadn't been fixed.

We've fixed those bugs in #2034 -- the "first" and "third" parts
of #1798 -- so now go ahead and offer the Refresh/Subscribe banner,
with the planned text, in the case where you can send messages.

This completes the work for #1798.