ci: Fetch all branches during the checkout step
See:
    https://github.com/actions/checkout/blob/8e8c483db/action.yml#L74-L76

Without this change, the main branch is not fetched when the
current branch is non-main. We need this because soon we will run
the `flutter_version` suite in CI only when there are changes in
`pubspec.yaml` or `tools/check` files. For that, we need the `main`
branch to be able to calculate the file diff.

Otherwise the suite will fail when calling `git_base_commit` saying:

    fatal: Not a valid object name refs/remotes/origin/main