ci: Run flutter_version suite conditionally
Flutter's new versioning depends on the latest tag pushed upstream.
And we are going to add the check for Flutter version string back
soon. To avoid frequent CI failures that may be caused by a new tag
being pushed upstream, run the flutter_version check in CI only
when there are changes in the `pubspec.yaml` or `tools/check` files.