compose [nfc]: Explain why quote-and-reply uses absolute URLs
In discussion on this PR, this related area looked to me like
something we should change, until I learned more:
  https://github.com/zulip/zulip-flutter/pull/1902#discussion_r2566957219
  https://github.com/zulip/zulip-flutter/pull/1902#discussion_r2604809054

Write that further context down, so that it's visible for the
next person who wonders the same thing.

Also explain why conversely we use relative URLs for the fallback
form of #-mention links.