changelog: Describe changes since 30.0.267
Trying a bit of a new format this time.  I'm hoping that by
consistently including the PR number for all changes, in addition
to the issue number where applicable, it'll be somewhat easier to
compare these against a Git log (decorated with PR refs) or a GitHub
list of merged PRs and closed issues.  That in turn should make it
easier to compile one of these -- and in particular to do so
incrementally, updating a partial version from one week to the next.

Relatedly, I think this form is likely to be more amenable to being
partly automated in the future.  Ideally the human would deal only
with identifying user-visible changes, describing them, and selecting
the user highlights.