notif: On Android handle notification taps via Pigeon API
Instead of relying on Flutter's deeplinks implementation for
routing the notification URL, handle the Android Intents generated
by notification taps ourselves using Pigeon to pass those events
over to the Dart layer from the Java layer.

The upstream Flutter's deeplinks implementation has a bug where if
the deeplink is triggered after the app was killed by the OS when
it was in background, the app will get launched again but the
route/link will not reach the Flutter's navigation handlers. See:
  https://github.com/flutter/flutter/issues/178305

In the failure case we seem to be receiving the Android Intent for
the notification tap from the OS via `MainActivity.onNewIntent`
without any problems. So, to workaround that upstream bug this
commit changes the implementation to handle these Android Intents
ourselves.

Fixes: #1567