check: Fix error condition in parsing Flutter version from pubspec
This is intended to check if the pattern failed to match, by
checking whether the array `parsed` has zero elements.  To do this,
it looks at the length of the array... but then it was checking if
that length, as a string, is empty.  Oops.  Instead, check if the
length, as a number, is zero.