compose: Add button to insert global time
This adds a clock button to the compose box that opens date and time
pickers sequentially. When completed, it inserts a timestamp in Zulip's
global time format: <time:2025-12-31T13:30:00+05:30>. This allows users
to easily reference specific times that display correctly in all timezones.

Fixes part of #1215.

Note: Localized strings for the picker dialogs will be updated in a
subsequent commit.