message test: Strengthen checks on when notifyListeners called
Generally we've taken care to ensure that the highest-frequency
events, including several events handled by MessageStore, don't
cause the overall PerAccountStore to notify its listeners.

The main effect of this test change is to verify that.  We already
have helpers checkNotifiedOnce and checkNotNotified, which numerous
tests in this file use to confirm that MessageListView either did or
didn't notify listeners in a given situation.  With this change,
those also verify that the PerAccountStore didn't notify listeners.

A secondary effect is that in the handful of spots (recently added)
where we do now expect PerAccountStore to notify listeners, we also
check whether MessageListView did so, just like we do in other tests.