notif [nfc]: Get current account from nav stack, not store
The behavior should be unchanged here, because the navigation observer
_UpdateLastVisitedAccount was already keeping the global store's
record of the current account up to date.

But logically the information this bit of code wants is about the
current navigation stack in the app's navigator.  So now that we
maintain an accessible record of that stack, let's inspect that
directly.